requirements.txt
absl-py                0.8.0
astor                  0.8.0
atari-py               0.2.6
atomicwrites           1.3.0
attrs                  19.1.0
awscli                 1.18.152
botocore               1.18.11
box2d-py               2.3.8
certifi                2019.9.11
cffi                   1.12.3
chardet                3.0.4
Click                  7.0
cloudpickle            1.2.2
colorama               0.4.1
cycler                 0.10.0
Cython                 0.29.13
decorator              4.4.2
dm-control             0.0.336440200
dm-env                 1.2
dm-tree                0.1.5
docutils               0.15.2
filelock               3.0.12
flatbuffers            1.11
funcsigs               1.0.2
future                 0.17.1
gast                   0.3.2
glfw                   1.8.3
grpcio                 1.23.0
gtimer                 1.0.0b5
gym                    0.15.7
h5py                   2.10.0
hyperopt               0.2.4
idna                   2.8
imageio                2.5.0
importlib-metadata     0.23
jmespath               0.10.0
joblib                 0.16.0
kiwisolver             1.1.0
labmaze                1.0.3
lockfile               0.12.2
lxml                   4.5.2
lz4                    2.2.1
Markdown               3.1.1
matplotlib             3.1.1
mjrl                   1.0.0
mock                   3.0.5
more-itertools         7.2.0
mpi4py                 3.0.3
mujoco-py              1.50.1.68
networkx               2.5
nibabel                2.1.0
nine                   1.1.0
numpy                  1.19.2
numpy-stl              2.2.0
opencv-python-headless 4.1.1.26
packaging              19.2
pandas                 1.1.0
path                   15.0.0
Pillow                 6.1.0
pip                    19.2.3
pluggy                 0.13.0
polling                0.3.1
progressbar2           3.53.1
protobuf               3.13.0
psutil                 5.6.3
py                     1.8.0
py-spy                 0.1.11
pyasn1                 0.4.8
pybullet               3.0.4
pycparser              2.19
pyglet                 1.4.4
pylru                  1.0.9
PyOpenGL               3.1.5
pyparsing              2.4.2
pytest                 5.1.2
python-dateutil        2.8.0
python-utils           2.4.0
pytz                   2020.1
PyYAML                 5.1.2
redis                  3.3.8
requests               2.22.0
rsa                    4.5
s3transfer             0.3.3
scipy                  1.3.1
seaborn                0.10.1
Serializable           0.1.0
setproctitle           1.1.10
setuptools             39.0.1
six                    1.12.0
svgwrite               1.4
tensorboard            1.13.1
termcolor              1.1.0
Theano                 1.0.5
torch                  1.3.0
torchvision            0.4.1
tqdm                   4.32.1
Tree                   0.2.4
typing                 3.7.4.1
urllib3                1.25.5
wcwidth                0.1.7
Werkzeug               0.16.0
wheel                  0.33.6
xmltodict              0.12.0
zipp                   0.6.0
